% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeMetadata.R
\name{make_metadata}
\alias{make_metadata}
\title{Generate the metadata.csv file from a documentation file}
\usage{
make_metadata(
  directory = character(0L),
  dataDirs = character(0L),
  version = "1.0.0",
  ext_pattern = character(0L),
  doc_file,
  pkg_name = character(0L),
  dry.run = TRUE,
  append = FALSE
)
}
\arguments{
\item{directory}{The base folder for \emph{all} datasets}

\item{dataDirs}{character() A vector of folder names contained in directory
that corresponds to each project. For multiple versions, repeat the
name of the folder.}

\item{version}{character() A vector of subfolder versions that is parallel
to \code{dataDirs} argument, typically \code{v1.0.0}.}

\item{ext_pattern}{character(1) A string that matches files within the
above folders to find the data.}

\item{doc_file}{character(1) A path to the documentation \code{data.frame} that
tells the function how to fill in the standard columns for data
annotation, for example \code{DataProvider}, \code{TaxonomyId}, etc.}

\item{pkg_name}{character(1) The name of the current package}

\item{dry.run}{logical(1) Whether to (over)write the \code{metadata.csv} file or
return as output.}

\item{append}{logical(1) Whether to append to the current \code{metadata.csv}
file}
}
\value{
Saves a file under \verb{/inst/extdata/metadata.csv}
}
\description{
This function takes a specific folder structure and generates the
metadata.csv file for adding to ExperimentHub.
}
\examples{

make_metadata(
    directory = character(0L),
    dataDirs = character(0L),
    version = "1.0.0",
    ext_pattern = character(0L),
    doc_file = "inst/extdata/docuData/CyCIFData_v1.csv",
    pkg_name = "CyCIFData",
    dry.run = TRUE
)

if (interactive()) {
    make_metadata(
        directory = "~/data/scmm",
        dataDirs = "pbmc",
        version = "1.0.0",
        ext_pattern = "\\\\.[Rr][Dd][AaSs]$|\\\\.[Mm][Tt][Xx]\\\\.[Gg][Zz]$",
        doc_file = "inst/extdata/docuData/singlecellmultimodalv6.csv",
        pkg_name = "SingleCellMultiModal",
        dry.run = TRUE,
    )
}

}
